+++ server/controller/admin/singer.controller.js
@@ -4,6 +4,14 @@ const moment = require("moment");
 const slugify = require("slugify")
 
 module.exports.createPost = async (req, res) => {
+  if (!req.account.permission.includes("singer-create")) {
+    res.json({
+      code: "error",
+      message: "Không có quyền sử dụng tính năng này!"
+    })
+    return;
+  }
+
   const numDocument = await Singer.countDocuments({});
   if (!req.body.position) {
     req.body.position = numDocument;
@@ -114,6 +122,13 @@ module.exports.applyMultiPatch = async (req, res) => {
   switch (req.body.status) {
     case "active": case "inactive":
       {
+        if (!req.account.permission.includes("singer-edit")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền sử dụng tính năng này!"
+          })
+          return;
+        }
         await Singer.updateMany({
           _id: { $in: req.body.idList }
         }, {
@@ -125,6 +140,13 @@ module.exports.applyMultiPatch = async (req, res) => {
       }
     case "delete":
       {
+        if (!req.account.permission.includes("singer-delete")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền sử dụng tính năng này!"
+          })
+          return;
+        }
         await Singer.updateMany({
           _id: { $in: req.body.idList }
         }, {
@@ -143,6 +165,13 @@ module.exports.applyMultiPatch = async (req, res) => {
 }
 
 module.exports.deletePatch = async (req, res) => {
+  if (!req.account.permission.includes("singer-delete")) {
+    res.json({
+      code: "error",
+      message: "Không có quyền sử dụng tính năng này!"
+    })
+    return;
+  }
   try {
     const id = req.params.id;
     await Singer.updateOne({
@@ -236,6 +265,13 @@ module.exports.trashListGet = async (req, res) => {
 }
 
 module.exports.trashApplyMultiPatch = async (req, res) => {
+  if (!req.account.permission.includes("singer-trash")) {
+    res.json({
+      code: "error",
+      message: "Không có quyền sử dụng tính năng này!"
+    })
+    return;
+  }
   switch (req.body.status) {
     case "hard-delete":
       {
@@ -264,7 +300,13 @@ module.exports.trashApplyMultiPatch = async (req, res) => {
 }
 
 module.exports.recoveryPatch = async (req, res) => {
-  try {
+  if (!req.account.permission.includes("singer-trash")) {
+    res.json({
+      code: "error",
+      message: "Không có quyền sử dụng tính năng này!"
+    })
+    return;
+  } try {
     const id = req.params.id;
     await Singer.updateOne({
       _id: id
@@ -287,6 +329,13 @@ module.exports.recoveryPatch = async (req, res) => {
 }
 
 module.exports.hardDelete = async (req, res) => {
+  if (!req.account.permission.includes("singer-trash")) {
+    res.json({
+      code: "error",
+      message: "Không có quyền sử dụng tính năng này!"
+    })
+    return;
+  }
   try {
     const id = req.params.id;
     await Singer.deleteOne({
@@ -333,6 +382,13 @@ module.exports.editGet = async (req, res) => {
 }
 
 module.exports.editPatch = async (req, res) => {
+  if (!req.account.permission.includes("singer-edit")) {
+    res.json({
+      code: "error",
+      message: "Không có quyền sử dụng tính năng này!"
+    })
+    return;
+  }
   try {
     const id = req.params.id;
